+++ app/controllers/secrets_controller.rb
@@ -10,6 +10,7 @@ class SecretsController < RestController
   include Secrets::RedisHandler
   include GroupMembershipValidator
   include EdgeValidator
+  include TriggerMessage
 
   before_action :current_user
 
@@ -18,7 +19,8 @@ class SecretsController < RestController
     if params[:action].downcase.starts_with?('show') || params[:action].downcase.starts_with?('batch')
       yield
     else
-      Sequel::Model.db.transaction(&block)
+      super(&block)
+      trigger_message_job
     end
   end
 
@@ -30,9 +32,7 @@ class SecretsController < RestController
     value = request.raw_post
 
     raise ArgumentError, "'value' may not be empty" if value.blank?
-
     ::DB::Service::SecretService.instance.secret_value_change(resource.id, value)
-    resource.enforce_secrets_version_limit
 
     head(:created)
   ensure