+++ app/models/pubsub/events/event_input.rb
@@ -7,6 +7,15 @@ class EventInput
     raise NotImplementedError
   end
 
+  # Generates event input based on specific implementation of `get_event_input` in derived class
+  # And creates event based on that input
+  def send_event(operation, db_obj)
+    if Rails.application.config.conjur_config.try(:conjur_pubsub_enabled)
+      event_type, event_value = get_event_input(operation, db_obj)
+      Event.create_event(event_type: event_type, event_value: event_value)
+    end
+  end
+
   protected
 
   def get_entity_type