+++ spec/lib/tasks/events_to_sns_spec.rb
@@ -1,0 +1,21 @@
+require 'spec_helper'
+Rails.application.load_tasks
+
+describe "events_to_sns:publish" do
+
+  before(:each) do
+    create_sns_topic
+    allow(Rails.application.config.conjur_config).to receive(:conjur_pubsub_enabled).and_return(true)
+  end
+
+  it "runs the MessageJob and logs errors if any" do
+    expect(MessageJob.instance).to receive(:run)
+    Event.create_event(event_type: 'test_event1', event_value: '{"key":"value1"}')
+    Rake::Task["events_to_sns:publish"].execute
+  end
+
+  it "logs an error if MessageJob fails" do
+    allow(MessageJob.instance).to receive(:run).and_raise(StandardError.new("Test error"))
+    Rake::Task["events_to_sns:publish"].invoke
+  end
+end